Refactor: Criação da pasta connections e ajuste em todas as rotas
