+++ src/controllers/client/listClient.js
@@ -1,4 +1,4 @@
-const knex = require('../../dbConnection');
+const knex = require('../../connections/dbConnection');
 
 
 const listClient = async (req, res) => {
