+++ src/middlewares/authentication.js
@@ -1,5 +1,5 @@
 const jwt = require('jsonwebtoken')
-const knex = require('../dbConnection')
+const knex = require('../../connections/dbConnection');
 
 const authentication = async (req, res, next) => {
   const { authorization } = req.headers
