+++ src/utils/checkFunctions/checkClient/checkClientCpfRegistered.js
@@ -1,4 +1,4 @@
-const knex = require('../../../dbConnection')
+const knex = require('../../connections/dbConnection');
 
 const checkCpfRegistered = async (cpf) => {
     const cpfExists = await knex('clientes').where({ cpf }).first()
