+++ src/utils/checkFunctions/checkClient/findClientById.js
@@ -1,4 +1,4 @@
-const knex = require('../../../dbConnection');
+const knex = require('../../connections/dbConnection');const knex = require('../../../dbConnection');
 
 
 const findClientById = async (id) => {
