+++ src/utils/checkFunctions/checkOrders/findClientByIDinOrdersTable.js
@@ -1,4 +1,4 @@
-const knex = require('../../../dbConnection');
+const knex = require('../../connections/dbConnection');
 
 
 const findClientId = async (id) => {
