+++ src/utils/checkFunctions/checkProducts/checkProductExistsIntoDb.js
@@ -1,4 +1,4 @@
-const knex = require('../../../dbConnection')
+const knex = require('../../connections/dbConnection');
 
 const checkProductExistsIntoDb= async (descricao) => {
   const productExists = await knex('produtos').where({ descricao }).first().returning('*')
