+++ src/utils/checkFunctions/checkProducts/findProductByID.js
@@ -1,4 +1,4 @@
-const knex = require('../../../dbConnection')
+const knex = require('../../connections/dbConnection');
 
 const findProduct = async (id) => {
   const product = await knex('produtos').where({
