+++ src/utils/insertFunctions/insertClientIntoDatabase.js
@@ -1,4 +1,4 @@
-const knex = require('../../dbConnection')
+const knex = require('../../connections/dbConnection');
 
 const insertClientIntoDatabase = async (nome, email, cpf, cep, rua, numero, bairro, cidade, estado) => {
     const client = await knex('clientes')
