+++ src/utils/insertFunctions/updateProductIntoDatabase.js
@@ -1,4 +1,4 @@
-const knex = require('../../dbConnection')
+const knex = require('../../connections/dbConnection');
 
 const updateProductIntoDatabase = async (id, descricao, quantidade_estoque, valor, categoria_id) => {
     const register = await knex('produtos')
